[FIX] 게시글관리 화면 표시 수정
